feat(order): JPA 연동, Repository 구현

+++ order/src/main/java/com/smart/commerce/order/module/order/application/dto/OrderRequest.java
@@ -1,4 +1,4 @@
 package com.smart.commerce.order.module.order.application.dto;
 
-public record OrderRequest(int userId, int storeId) {
+public record OrderRequest(long userId, int storeId) {
 }
